Rename XampleApp to WorldLocking.ASA.Examples for consistency with main repo.
