+++ Advanced/ASA/Assets/WorldLocking.ASA.Examples/Materials/Xample_Wireframe.mat.meta
