+++ Advanced/ASA/Assets/WorldLocking.ASA.Examples/Microsoft.MixedReality.WorldLocking.ASA.Examples.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3fad2220a89bf9c4fb7576337c8b8d55
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
