+++ Advanced/ASA/Assets/WorldLocking.ASA.Examples/Prefabs/Shafter.prefab.meta
