+++ Advanced/ASA/Assets/WorldLocking.ASA.Examples/Prefabs/ShafterWhite.prefab
