+++ Advanced/ASA/Assets/WorldLocking.ASA.Examples/Scripts/ActionPublish.cs
@@ -10,8 +10,11 @@ using Microsoft.MixedReality.WorldLocking.ASA;
 
 using TMPro;
 
-namespace WorldLocking.XampleApp
+namespace Microsoft.MixedReality.WorldLocking.ASA.Examples
 {
+    /// <summary>
+    /// Script for implementing button actions. Simple passthrough to perform publisher actions on interaction events.
+    /// </summary>
     public class ActionPublish : ActionCube
     {
         public SpacePinBinder spacePinBinder;
@@ -87,12 +90,12 @@ namespace WorldLocking.XampleApp
         {
             if (statusLine != null)
             {
-                var status = new IPublisher.ReadinessStatus();
+                var status = new ReadinessStatus();
                 if (binder != null)
                 {
                     status = binder.PublisherStatus;
                 }
-                statusLine.faceColor = status.readiness == IPublisher.Readiness.Ready ? Color.white : Color.red;
+                statusLine.faceColor = status.readiness == PublisherReadiness.Ready ? Color.white : Color.red;
                 statusLine.text = $"Status: {status.readiness.ToString()}, Create={status.recommendedForCreate.ToString("0.00")}, {status.readyForCreate.ToString("0.00")}";
             }
         }
@@ -260,3 +263,4 @@ namespace WorldLocking.XampleApp
         }
     }
 }
+
